+++ hknweb/studentservices/admin.py
@@ -1,8 +1,9 @@
 from django.contrib import admin
 
-from hknweb.studentservices.models import Resume
+from hknweb.studentservices.models import Resume, ReviewSession
 
 
+@admin.register(Resume)
 class ResumeAdmin(admin.ModelAdmin):
     fields = ["name", "email", "notes", "document", "uploaded_at", "critiques"]
     readonly_fields = ["uploaded_at", "email", "name", "document"]
@@ -10,4 +11,28 @@ class ResumeAdmin(admin.ModelAdmin):
     list_filter = ("name", "notes", "document", "uploaded_at")
 
 
-admin.site.register(Resume, ResumeAdmin)
+@admin.register(ReviewSession)
+class ReviewSessionAdmin(admin.ModelAdmin):
+
+    fields = [
+        "name",
+        "slug",
+        "start_time",
+        "end_time",
+        "location",
+        "description",
+        "created_by",
+        "created_at",
+    ]
+    # NOTE: created_by should be read only, but I don't know how to set it to default to current user
+    readonly_fields = ["created_at"]
+    list_display = ("name", "start_time", "location", "created_by", "created_at")
+    list_filter = ["start_time", "created_at", "location", "created_by"]
+    search_fields = [
+        "name",
+        "created_by__username",
+        "created_by__first_name",
+        "created_by__last_name",
+    ]
+    ordering = ["-created_at"]
+    autocomplete_fields = ["created_by"]
