+++ hknweb/studentservices/templates/studentservices/reviewsession_edit.html
@@ -1,6 +1,6 @@
 {% extends "base.html" %}
 {% load static %}
-{% block title %}Add a review session{% endblock %}
+{% block title %}Edit {{ reviewsession.name }}{% endblock %}
 
 {% block header %}
 <!-- jQuery -->
@@ -10,7 +10,7 @@
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jquery-datetimepicker/2.5.20/jquery.datetimepicker.min.css" integrity="sha256-DOS9W6NR+NFe1fUhEE0PGKY/fubbUCnOfTje2JMDw3Y=" crossorigin="anonymous" />
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-datetimepicker/2.5.20/jquery.datetimepicker.full.min.js" integrity="sha256-FEqEelWI3WouFOo2VWP/uJfs1y8KJ++FLh2Lbqc8SJk=" crossorigin="anonymous"></script>
 
-<link rel="stylesheet" href="{% static "reviewsessions/style.css" %}">
+<link rel="stylesheet" href="{% static "studentservices/style.css" %}">
 <style>
     .heading {
         display: none;
@@ -32,35 +32,22 @@
 </style>
 {% endblock %}
 
-{% block heading %}Add a review session{% endblock %}
+{% block heading %}Edit Review Session: {{ reviewsession.name }}{% endblock %}
 
 {% block content %}
-<h1 id="add-review-session-title"> Add a review session </h1>
+<h1 id="add-review-session-title"> Edit Review Session: {{ reviewsession.name }} </h1>
 
 <div class="centered">
-    <form method="POST" id="form" onsubmit="submit_button.disabled = true; return true;">
+    <form method="post" onsubmit="submit_button.disabled = true; return true;">
         {% csrf_token %}
-        {% for field in form %}
-            <div class="row">
-                <span class="left-side">
-                    {{ field.label }}
-                    {% if field.field.required %}
-                        <span class="message">*</span>
-                    {% endif %}
-                </span>
-                {{ field }}
-                {{ field.errors }}
-            </div>
-        {% endfor %}
-        <div id="submit">
-            <input type="submit" name="submit_button" value="Submit" />
-            {% if messages %}
-                {% for message in messages %}
-                    <br>
-                    <span class="message">{{ message }}</span>
-                {% endfor %}
-            {% endif %}
-        </div>
+        {{ form.as_p }}
+        <input type="submit" name="submit_button" value="Update">
+        {% if messages %}
+            {% for message in messages %}
+                <br>
+                <span class="message">{{ message }}</span>
+            {% endfor %}
+        {% endif %}
     </form>
 
     <script>
