+++ hknweb/templates/base.html
@@ -99,7 +99,7 @@
       <span class="slash">/</span>
       <a href="/tutoring">Tutoring Hours</a>
       <span class="slash">/</span>
-      <a href="/reviewsessions">Review Sessions</a>
+      <a href="/studentservices/reviewsessions">Review Sessions</a>
       <span class="slash">/</span>
       <a href="{% url 'course_surveys:index' %}">Course Surveys</a>
       <span class="slash">/</span>
