+++ hknweb/templates/landing/home.html
@@ -153,7 +153,7 @@
             <ul>
                 {% for reviewsession in reviewsessions %}
                     <li>
-                        <a href="{% url 'reviewsessions:index' %}{{ reviewsession.id }}">{{ reviewsession.name }}</a>
+                        <a href="{% url 'studentservices:reviewsessions' %}{{ reviewsession.get_absolute_url }}">{{ reviewsession.name }}</a>
                         <span class="events-card-time">{{reviewsession.start_time|date:"D m/d h:i A"}} - {{ reviewsession.end_time|date:"h:i A"}}</span>
                     </li>
                 {% endfor %}
